Fix use of raised falcon errors to use 2 params
Refactoring included to clean up patterns of usage and improve
readability of the logic.  In the process added support for standard
form media type handling as an alternative to json for this API as
it would be sufficient in this case.

Noting this repo needs an api schema, tests, and more.